Unify matrix for split common and regular matrix
